+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### 0.5.3 (Upcoming release)
 - Add tests for GreyNumber type
-- topsis() and aras() accepts matrix
+- methods now accept Matrix (in addition to DataFrame)
+
 
 ### 0.5.2 
 - Implement MEREC